+++ src/dgram/unicast.ts
@@ -41,9 +41,9 @@ export class Unicast extends Dgram {
    *
    * @param {string} name - The name of the socket.
    * @param {'udp4' | 'udp6'} socketType - The type of the socket (IPv4 or IPv6).
-   * @param {boolean | undefined} reuseAddr - Whether to allow address reuse.
-   * @param {string} [interfaceName] - The name of the network interface to bind to.
-   * @param {string} [interfaceAddress] - The address of the network interface to bind to.
+   * @param {boolean | undefined} reuseAddr - Whether to allow address reuse. Defaults to true.
+   * @param {string} [interfaceName] - The name of the network interface to bind to. If provided, the interfaceAddress will be determined based on this name if not explicitly provided.
+   * @param {string} [interfaceAddress] - The address of the network interface to bind to. If not provided, it will be determined based on the interfaceName.
    * @param {number} [port] - The port number to bind to. If not provided, it will bind to any available port.
    */
   constructor(name: string, socketType: 'udp4' | 'udp6', reuseAddr: boolean | undefined = true, interfaceName?: string, interfaceAddress?: string, port?: number) {